BL-1815 Reorganize helper methods
