+++ spec/helpers/catalog_helper_spec.rb
@@ -84,164 +84,6 @@ RSpec.describe CatalogHelper, type: :helper do
     end
   end
 
-  describe "#render_purchase_order_availability" do
-    let(:user) { FactoryBot.build(:user) }
-    let(:doc) { SolrDocument.new(purchase_order: true, id: "foo") }
-    let(:can_purchase_order?) { true }
-    let(:config) { CatalogController.blacklight_config }
-    let(:context) { Blacklight::Configuration::Context.new(config) }
-    let(:presenter) { helper.document_presenter(doc) }
-
-    before(:each) do
-      allow(helper).to receive(:link_to) { "render_login_link" }
-      allow(helper).to receive(:current_user) { user }
-      allow(helper).to receive(:content_tag) {}
-      allow(helper).to receive(:render) {}
-      allow(user).to receive(:can_purchase_order?) { can_purchase_order? }
-
-      without_partial_double_verification do
-        allow(helper).to receive(:blacklight_config) { config }
-        allow(helper).to receive(:blacklight_configuration_context) { context }
-      end
-
-      helper.render_purchase_order_availability(presenter)
-    end
-
-    context "document has purchase order and user is not logged in" do
-      let(:user) { nil }
-
-      it "should render the log_in link" do
-        expect(helper).to have_received(:render).with(
-          partial: "purchase_order_anonymous_button",
-          locals: { document: doc, link: "render_login_link" }
-        )
-      end
-    end
-
-    context "document has purchase order and user is not logged in and link configured to appear in button" do
-      let(:user) { nil }
-      let(:args) { {
-        document: SolrDocument.new(purchase_order: true, id: "foo"),
-        config: { with_po_link: true },
-      } }
-
-      it "should render the log_in link inside of button" do
-        expect(helper).to have_received(:render).with(
-          partial: "purchase_order_anonymous_button",
-          locals: { document: args[:document], link: "render_login_link" }
-        )
-      end
-    end
-
-    context "document has purchase order and user is logged in" do
-      it "should render the purchase order button" do
-        expect(helper).to have_received(:content_tag).with(
-          :div, "render_login_link", class: "requests-container mb-2 ml-0"
-        )
-      end
-    end
-
-    context "document has purchase order but user cannot purchase order" do
-      let(:can_purchase_order?) { false }
-
-      it "should render purchase allow message" do
-        expect(helper).to have_received(:content_tag).with(
-          :div, t("purchase_order.purchase_order_allowed"), class: "availability"
-        )
-      end
-    end
-
-    context "document does not have purchase order button" do
-      let(:doc) { SolrDocument.new(purchase_order: false) }
-
-      it "should not render the purchase_order_button" do
-        expect(helper.render_purchase_order_availability(presenter)).to be_nil
-      end
-    end
-  end
-
-  describe "#render_purchase_order_show_link" do
-    let(:args) { { document: SolrDocument.new(purchase_order: true, id: "foo") } }
-    let(:user) { FactoryBot.build_stubbed(:user) }
-    let(:can_purchase_order?) { true }
-
-    before(:each) do
-      allow(helper).to receive(:link_to) { "render_login_link" }
-      allow(helper).to receive(:current_user) { user }
-      allow(helper).to receive(:render_purchase_order_button) { "render_purchase_order_button" }
-      allow(user).to receive(:can_purchase_order?) { can_purchase_order? }
-
-      without_partial_double_verification do
-        allow(helper).to receive(:blacklight_config) { blacklight_config }
-      end
-
-      helper.render_purchase_order_show_link(args)
-    end
-
-    context "document has purchase order and user is not logged in" do
-      let(:user) { nil }
-
-      it "should render the log in in link" do
-        expect(helper.render_purchase_order_show_link(args)).to eq("render_login_link")
-      end
-    end
-
-    context "document does not have purchase order" do
-      let(:args) { { document: SolrDocument.new(purchase_order: false, id: "foo") } }
-
-      it "should not render the log in in link" do
-        expect(helper.render_purchase_order_show_link(args)).to be_nil
-      end
-    end
-
-    context "user is logged in and cannot purchase an order" do
-      let(:can_purchase_order?) { false }
-
-      it "should not render the log in in link" do
-        expect(helper.render_purchase_order_show_link(args)).to be_nil
-      end
-    end
-
-    context "user is logged in and can purchase an order" do
-      it "should not render the log in in link" do
-        expect(helper.render_purchase_order_show_link(args)).to eq("render_purchase_order_button")
-      end
-    end
-  end
-
-  describe "#render_alma_availability(document)" do
-    let(:doc) { SolrDocument.new(bound_with_ids: ["foo"]) }
-    let(:config) { CatalogController.blacklight_config }
-
-    before do
-      without_partial_double_verification do
-        allow(helper).to receive(:blacklight_config) { config }
-      end
-    end
-
-    context "with bound_with_ids defined" do
-      it "renders the bound_with_ids" do
-        expect(helper.render_alma_availability(doc)).not_to be_nil
-      end
-    end
-
-    context "with no bound with ids available" do
-      let(:doc) { SolrDocument.new(bound_with_ids: nil) }
-
-      it "does not render the bound_with_ids" do
-        expect(helper.render_alma_availability(doc)).to be_nil
-      end
-    end
-
-    context "without bound_with_ids configured" do
-      let(:config) { PrimoCentralController.blacklight_config }
-
-      it "does not render the bound_with_ids" do
-        expect(helper.render_alma_availability(doc)).to be_nil
-      end
-    end
-  end
-
   describe "#render_email_form_field" do
     let(:current_user) { OpenStruct.new(email: nil) }
 
@@ -265,33 +107,6 @@ RSpec.describe CatalogHelper, type: :helper do
     end
   end
 
-  describe "#get_unavailable_notes" do
-    let(:service_notes) {  { "foo" => { "value" => "foo" } } }
-
-    before do
-      allow(helper).to receive(:electronic_notes).with("service") { service_notes }
-    end
-
-    context "with no unavailable notes" do
-      it "should not return any unavailability notes" do
-        expect(helper.get_unavailable_notes("bizz")).to eq([])
-      end
-    end
-
-    context "with unavailable notes" do
-      let(:service_notes) {  { "bizz" => {
-        "key" => "foo",
-        "service_temporarily_unavailable" => "foo",
-        "service_unavailable_reason" => "bar",
-        "service_unavailable_date" => "buzz",
-      } } }
-
-      it "should not return any unavailability notes" do
-        expect(helper.get_unavailable_notes("bizz")).to eq(["This service is temporarily unavailable due to: bar."])
-      end
-    end
-  end
-
   describe "#get_search_params" do
     context "with title_uniform_display field" do
       let(:field) { "title_uniform_display" }
@@ -309,223 +124,6 @@ RSpec.describe CatalogHelper, type: :helper do
 
       it "should not add quotes to the query or use title search" do
         expect(get_search_params(field, query)).to eq({ q: "Beethoven, Symphony no. 5", search_field: "title_statement_vern_display" })
-      end
-    end
-  end
-
-  describe "#render_electronic_notes" do
-    let(:service_notes) {  { "foo" => { "value" => "foo" } } }
-    let(:collection_notes) {  { "bizz" => { "value" => "bar" } } }
-    let(:public_notes) { "public note" }
-
-    before do
-      allow(helper).to receive(:render) { "rendered note" }
-      allow(Rails.cache).to receive(:fetch).with("collection_notes") { collection_notes }
-      allow(Rails.cache).to receive(:fetch).with("service_notes") { service_notes }
-    end
-
-    context "with no notes" do
-      let(:field) { {} }
-
-      it "should not render any notes" do
-        expect(render_electronic_notes(field)).to be_nil
-      end
-    end
-
-    context "with public notes" do
-      let(:field) { { "public_note" => "public note" } }
-
-      it "should render the notes" do
-        expect(render_electronic_notes(field)).to eq("rendered note")
-      end
-    end
-
-    context "with authentication notes" do
-      let(:field) { { "authentication_note" => "authentication note" } }
-
-      it "should render the notes" do
-        expect(render_electronic_notes(field)).to eq("rendered note")
-      end
-
-      it "should parse the authenticated notes" do
-        allow(helper).to receive(:render) do |arg|
-          expect(arg.dig(:locals, :authentication_notes)).to eq("authentication note")
-          ""
-        end
-
-        render_electronic_notes(field)
-      end
-    end
-
-    context "with service notes" do
-      let(:field) { { "service_id" => "foo" } }
-
-      it "should render the notes" do
-        expect(render_electronic_notes(field)).to eq("rendered note")
-      end
-    end
-
-    context "with collection notes" do
-      let(:field) { { "collection_id" => "bizz" } }
-
-      it "should render the notes" do
-        expect(render_electronic_notes(field)).to eq("rendered note")
-      end
-    end
-
-    context "with both collection and service notes" do
-      let(:field) { { "service_id" => "foo", "collection_id" => "bizz" } }
-
-      it "should render the notes" do
-        expect(render_electronic_notes(field)).to eq("rendered note")
-      end
-    end
-
-    context "with unavailable notes" do
-      let(:field) { { "service_id" => "buzz", "collection_id" => "bizz" } }
-      let(:service_notes) { { "foo" => {
-        "service_temporarily_unavailable" => "foo",
-        "service_unavailable_date" => "bar",
-        "service_unavailable_reason" => "buzz"
-      } } }
-
-      it "should render the notes" do
-        expect(render_electronic_notes(field)).to eq("rendered note")
-      end
-    end
-  end
-
-  describe "#electronic_resource_link_builder(field)" do
-    let(:alma_domain) { "sandbox01-na.alma.exlibrisgroup.com" }
-    let(:alma_institution_code) { "01TULI_INST" }
-
-    context "only a portfolio_pid is present" do
-      let(:field) { { "portfolio_id" => "12345" } }
-
-      it "has correct link to resource" do
-        expect(electronic_resource_link_builder(field)).to have_link(text: "Find it online", href: "https://sandbox01-na.alma.exlibrisgroup.com/view/uresolver/01TULI_INST/openurl?Force_direct=true&portfolio_pid=12345&rfr_id=info%3Asid%2Fprimo.exlibrisgroup.com&u.ignore_date_coverage=true")
-      end
-
-      it "does not have a separator" do
-        expect(electronic_resource_link_builder(field)).to_not have_text(" - ")
-      end
-    end
-
-    context "porfolio_id and collection name are present" do
-      let(:field) { { "portfolio_id" => "77777", "title" => "Sample Name" } }
-
-      it "displays database name if available" do
-        expect(electronic_resource_link_builder(field)).to have_link(text: "Sample Name", href: "https://sandbox01-na.alma.exlibrisgroup.com/view/uresolver/01TULI_INST/openurl?Force_direct=true&portfolio_pid=77777&rfr_id=info%3Asid%2Fprimo.exlibrisgroup.com&u.ignore_date_coverage=true")
-      end
-
-      it "does not contain a separator" do
-        expect(electronic_resource_link_builder(field)).to_not have_text(" - ")
-      end
-    end
-
-    context "only electronic note and collection name are present" do
-      let(:field) { { "portfolio_id" => "77777", "title" => "Sample Name" } }
-
-      it "does not contain a separator" do
-        allow(helper).to receive(:render_electronic_notes) { "Hello World" }
-        expect(electronic_resource_link_builder(field)).to_not have_text(" - ")
-      end
-    end
-
-    context "porfolio_id, collection name, and coverage statement are present" do
-      let(:field) { {
-        "portfolio_id" => "77777", "title" => "Sample Name", "coverage_statement" => "Sample Text"
-      } }
-
-      it "displays additional information as plain text" do
-        expect(electronic_resource_link_builder(field)).to have_link(text: "Sample Text", href: "https://sandbox01-na.alma.exlibrisgroup.com/view/uresolver/01TULI_INST/openurl?Force_direct=true&portfolio_pid=77777&rfr_id=info%3Asid%2Fprimo.exlibrisgroup.com&u.ignore_date_coverage=true")
-        expect(electronic_resource_link_builder(field)).to have_text("Sample Name")
-      end
-    end
-
-    context "porfolio_id, collection name, and coverage statement is not present present" do
-      let(:field) { {
-        "portfolio_id" => "77777", "title" => "Sample Name"
-      } }
-
-      it "displays additional information as plain text" do
-        expect(electronic_resource_link_builder(field)).to have_link(text: "Sample Name", href: "https://sandbox01-na.alma.exlibrisgroup.com/view/uresolver/01TULI_INST/openurl?Force_direct=true&portfolio_pid=77777&rfr_id=info%3Asid%2Fprimo.exlibrisgroup.com&u.ignore_date_coverage=true")
-      end
-    end
-
-    context "item is not available" do
-      let(:field) { { "availability" => "Not Available" } }
-
-      it "skips items that are not available" do
-        expect(electronic_resource_link_builder(field)).to be_nil
-      end
-    end
-  end
-
-  describe "#check_for_full_http_link(args)" do
-    let(:alma_domain) { "sandbox01-na.alma.exlibrisgroup.com" }
-    let(:alma_institution_code) { "01TULI_INST" }
-    let(:args) {
-        {
-          document:
-          {
-            electronic_resource_display: [
-              { "title" => "Access electronic resource.", "url" => "http://libproxy.temple.edu/login?url=http://www.aspresolver.com/aspresolver.asp?SHM2;1772483" },
-              { "portfolio_id" => "77777", "title" => "Sample Name" },
-            ]
-          },
-          field: :electronic_resource_display
-        }
-      }
-
-    context "marc field links with http should use the electronic_access_links helper method" do
-      it "directs an http link through the electronic_access_links method" do
-        expect(check_for_full_http_link(args)).to have_link(text: "Access electronic resource", href: "http://libproxy.temple.edu/login?url=http://www.aspresolver.com/aspresolver.asp?SHM2;1772483")
-      end
-    end
-
-    context "alma api links go through electronic_resource_display method" do
-      it "directs an http link through the electronic_access_links method" do
-        expect(check_for_full_http_link(args)).to have_link(text: "Sample Name", href: "https://sandbox01-na.alma.exlibrisgroup.com/view/uresolver/01TULI_INST/openurl?Force_direct=true&portfolio_pid=77777&rfr_id=info%3Asid%2Fprimo.exlibrisgroup.com&u.ignore_date_coverage=true")
-      end
-    end
-  end
-
-  describe "#electronic_access_links(field)" do
-    context "with only a url" do
-      let(:field) { { "url" => "http://libproxy.temple.edu/login?url=http://www.aspresolver.com/aspresolver.asp?SHM2;1772483" } }
-
-      it "has generic message for link" do
-        expect(electronic_access_links(field)).to have_link(text: "Link to Resource", href: "http://libproxy.temple.edu/login?url=http://www.aspresolver.com/aspresolver.asp?SHM2;1772483")
-      end
-    end
-
-    context "with collection name and url" do
-      let(:field) { { "title" => "Access electronic resource.", "url" => "http://libproxy.temple.edu/login?url=http://www.aspresolver.com/aspresolver.asp?SHM2;1772483" } }
-
-      it "displays z3 subfields if available" do
-        expect(electronic_access_links(field)).to have_link(text: "Access electronic resource", href: "http://libproxy.temple.edu/login?url=http://www.aspresolver.com/aspresolver.asp?SHM2;1772483")
-      end
-    end
-  end
-
-  describe "#single_link_builder(field)" do
-    let(:alma_domain) { "sandbox01-na.alma.exlibrisgroup.com" }
-    let(:alma_institution_code) { "01TULI_INST" }
-
-    context "with a url" do
-      let(:field) { { "url" => "http://libproxy.temple.edu/login?url=http://www.aspresolver.com/aspresolver.asp?SHM2;1772483" } }
-
-      it "single link is url" do
-        expect(single_link_builder(field)).to eq("http://libproxy.temple.edu/login?url=http://www.aspresolver.com/aspresolver.asp?SHM2;1772483")
-      end
-    end
-
-    context "without a url" do
-      let(:field) { { "portfolio_id" => "53395029150003811" } }
-
-      it "has generic message for link" do
-        expect(single_link_builder(field)).to eq("https://sandbox01-na.alma.exlibrisgroup.com/view/uresolver/01TULI_INST/openurl?Force_direct=true&portfolio_pid=53395029150003811&rfr_id=info%3Asid%2Fprimo.exlibrisgroup.com&u.ignore_date_coverage=true")
       end
     end
   end
@@ -906,63 +504,6 @@ end
         }
       it "returns false" do
         expect(open_shelves_allowed?(document)).to be false
-      end
-    end
-  end
-
-  describe "#build_hathitrust_url(field)" do
-    let(:field) { { "bib_key" => "000005117", "access" => "allow" } }
-    let(:base_url) { "https://catalog.hathitrust.org/Record/000005117?signon=swle:https://fim.temple.edu/idp/shibboleth" }
-    let(:constructed_url) { helper.build_hathitrust_url(field) }
-
-    it "returns a correctly formed url" do
-      expect(constructed_url).to eq base_url
-    end
-  end
-
-  describe "#hathitrust_link_allowed?(document))" do
-    context "record has a hathi_trust_bib_key_display field" do
-      context "with allow access" do
-        let(:document) { { "hathi_trust_bib_key_display" => [ { "bib_key" => "000005117", "access" => "allow" } ] } }
-
-        it "returns true" do
-          expect(hathitrust_link_allowed?(document)).to be(true)
-        end
-      end
-
-      context "with deny access" do
-        let(:document) { { "hathi_trust_bib_key_display" => [ { "bib_key" => "000005117", "access" => "deny" }] } }
-
-        it "does not render the online partial" do
-          expect(hathitrust_link_allowed?(document)).to be(false)
-        end
-      end
-    end
-  end
-
-  describe "#render_hathitrust_display(document)" do
-    context "record has a hathi_trust_bib_key_display field" do
-      context "with allow access" do
-        let(:document) { { "hathi_trust_bib_key_display" => [ { "bib_key" => "000005117", "access" => "allow" } ] } }
-
-        it "renders the online partial" do
-          expect(helper.render_hathitrust_display(document)).not_to be_nil
-        end
-      end
-
-      context "with deny access" do
-        let(:document) { { "hathi_trust_bib_key_display" => [ { "bib_key" => "000005117", "access" => "deny" }] } }
-
-        it "does not render the online partial" do
-          expect(helper.render_hathitrust_display(document)).to be_nil
-        end
-
-        context "when campus closed flag is true" do
-          it "renders the online partial" do
-            allow(helper).to receive(:campus_closed?).and_return("true")
-            expect(helper.render_hathitrust_display(document)).not_to be_nil
-          end
-        end
       end
     end
   end
